+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_02_08_220000) do
+ActiveRecord::Schema[8.1].define(version: 2026_02_11_100000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -79,6 +79,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_02_08_220000) do
     t.text "contact_info", null: false
     t.datetime "created_at", null: false
     t.text "description", null: false
+    t.string "era", default: "current", null: false
     t.string "part", null: false
     t.decimal "price", precision: 10, scale: 2, null: false
     t.datetime "updated_at", null: false